Wait until all Promises are settled before letting test fail
Previously, regression tests awaited Promise.all(), which rejected as
soon as any of the Promises rejected.
This meant that as far as Jest was concerned, all the tests were
finished, while in reality some of the Promises had yet to settle.
This led to snapshots being incorrectly diagnosed as obsolete, and
snapshots being written to the wrong test scope.
After this commit, if one of the Promises fails, the test still fail,
but won't fail until all the other Promises have settled.